chore: move Dimensions to the domain
